Update recent activities to display related vals
Updated recent activities to display only information if you are related
to the event. Participating on an event should display participation
with volunteering hours. Expenses that are reported will also be added
in the recent activities list container. Items that are added will also
be personally notified with the Notification tab under profile page.